// plugins/nx-prisma/src/executors/seed/schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Seed your database
 */
export interface SeedExecutorSchema {
  /**
   * The path to the seeding script
   */
  script: string;
  /**
   * TypeScript config to use for seeding
   */
  tsConfig?: string;
  /**
   * Tool to use for executing the seeding script. Currently `ts-node` and `tsx` are supported
   */
  executeWith?: 'ts-node' | 'tsx';
}